Add Oribos link to commodity shortages
