+++ app/routes/wow/full-scan/getOribosLink.tsx
@@ -1,0 +1,27 @@
+import ExternalLink from '~/components/utilities/ExternalLink'
+
+const parseServerName = (serverName: string) =>
+  serverName
+    .replaceAll('-', '')
+    .replaceAll("'", '')
+    .replaceAll(' ', '-')
+    .toLowerCase()
+
+export const getOribosLink =
+  (serverName: string | undefined, title: string) =>
+  ({ row }: { row: { itemID: number } }) => {
+    const itemId = row.itemID
+    if (typeof itemId !== 'number') return null
+
+    if (!serverName) return null
+
+    const parsedServerName = parseServerName(serverName)
+
+    return (
+      <ExternalLink
+        link={`https://oribos.exchange/#us-${parsedServerName}/${itemId}`}
+        text={title}
+        tooltip={`Oribos Marketplace For ${serverName}`}
+      />
+    )
+  }
